% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjTabone.R
\name{my_quantile}
\alias{my_quantile}
\title{A helper function to pivot confidence intervals for normal/non-normal variables.}
\usage{
my_quantile(x, probs)
}
\description{
A helper function to pivot confidence intervals for normal/non-normal variables.
}
