% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjTabone.R
\name{run_anova}
\alias{run_anova}
\title{This functions runs anova on normally distributed columns/variables. It's used by summarize_normal.}
\usage{
run_anova(data, strata, i)
}
\arguments{
\item{data}{Dataframe of your data.}

\item{strata}{The column to use for stratification of the data into cases/controls.}

\item{i}{The normal column/variable that is being stratified.}
}
\description{
This functions runs anova on normally distributed columns/variables. It's used by summarize_normal.
}
