% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjTabone.R
\name{run_fisher}
\alias{run_fisher}
\title{Runs the Fisher exact test and returns a dataframe with p-values. This function is used by summarize_categorical().}
\usage{
run_fisher(data, strata, i, idcol)
}
\arguments{
\item{data}{A dataframe with data.}

\item{strata}{The column to use for stratification of the data into cases/controls.}

\item{i}{The categorical column/variable that is being stratified.}

\item{idcol}{This is the column with unique identifiers for each row of the data. This is the column that will be used to generate the counts to run the test.}
}
\description{
Runs the Fisher exact test and returns a dataframe with p-values. This function is used by summarize_categorical().
}
