% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjTabone.R
\name{run_kwtest}
\alias{run_kwtest}
\title{This function runs the Kruskal-Wallis test on non-normal data. This function is used by summarize_nonnormal.}
\usage{
run_kwtest(data, strata, i)
}
\arguments{
\item{data}{Dataframe of your data.}

\item{strata}{The column to use for stratification of the data into cases/controls.}

\item{i}{The non-normal column/variable that is being stratified.}
}
\description{
This function runs the Kruskal-Wallis test on non-normal data. This function is used by summarize_nonnormal.
}
