% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjTabone.R
\name{sjTabone}
\alias{sjTabone}
\title{Summarize your dataset and and compare between stratifying groups.}
\usage{
sjTabone(data, myvars, idcol, strata, grps = NULL)
}
\arguments{
\item{data}{Dataframe of your data.}

\item{myvars}{Vector of column names form your dataset that need to be summarized. Categorical variables wil need to be cast as factors before passing the dataframe to the function.}

\item{idcol}{This is the column with unique identifiers for each row of the data. This is the column that will be used to generate the counts to run the test.}

\item{strata}{The column to use for stratification of the data into cases/controls.}

\item{grps}{If your stratifying column has more than 2 groups, and you only want to summarize a subset of them, pass those values to this function.}
}
\description{
Function to summarize your dataset. For every variable in the vector myvars the function checks whether it's categorical, or continuous.
If it's continuous it tests the normality of the variable to summarize appropriately.
It returns a dataframe with counts and percentages for the categorical variable. The normally distributed variables are summarized by mean and standard deviation.
Non-normal variables are summarized median and inter-quartile ranges.
Statistical comparisons are made between the different groups in the stratification column using the aprropriate statistical tests.
Categorical variables are compared using the Fisher exact test.
Normal variables are compared using ANOVA.
Non-normal variables are compared using the Kruskal-Wallis test.
}
