% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjTabone.R
\name{summarize_categorical}
\alias{summarize_categorical}
\title{Summarize Categorical variables.}
\usage{
summarize_categorical(data, strata, i, idcol, grps = NULL)
}
\arguments{
\item{data}{Dataframe of your data.}

\item{strata}{The column to use for stratification of the data into cases/controls.}

\item{i}{The categorical column/variable that is being stratified.}

\item{idcol}{This is the column with unique identifiers for each row of the data. This is the column that will be used to generate the counts to run the test.}

\item{grps}{If your stratifying column has more than 2 groups, and you only want to summarize a subset of them, pass those values to this function.}
}
\description{
Function to summarize categorical variables. It's used by sjtabone.
This function can also be used as a stand alone function.
}
