% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjTabone.R
\name{summarize_normal}
\alias{summarize_normal}
\title{Function to summarize normal variables.}
\usage{
summarize_normal(data, strata, i, grps = NULL)
}
\arguments{
\item{data}{Dataframe of your data.}

\item{strata}{The column to use for stratification of the data into cases/controls.}

\item{i}{The normal column/variable that is being stratified.}

\item{grps}{If your stratifying column has more than 2 groups, and you only want to summarize a subset of them, pass those values to this function.}
}
\description{
This function is used by sjtabone. It can also be used as a stand alone function.
}
