% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjTabone.R
\name{write_tabone}
\alias{write_tabone}
\title{Write the output of the sjtabone function to a word document.}
\usage{
write_tabone(tab.one, filename)
}
\arguments{
\item{tab.one}{A dataframe output from the function sjTabone.}

\item{filename}{Filename to write the output. It automatically adds a docx extension}
}
\description{
Write the output of the sjtabone function to a word document.
}
